CareChainAI/backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0          # PostgreSQL async driver (production)
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Document processing ───────────────────────────────────────────────────────
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0        # scanned PDF → images for OCR

# ── AI & Embeddings ───────────────────────────────────────────────────────────
google-generativeai==0.7.2
sentence-transformers==3.1.0
faiss-cpu==1.8.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2       # QR code generation
